[#103] Refactor : DhcBottomBar, DhcBottomBarState 모두 designsystem 모듈로 이동
